esp: Implement Wi-Fi connection manager (#4) (#12)
When the device is not connected to a network, the manager will:
1. Try to reconnect to the previous network, if it was saved
2. Try to reconnect to in-range saved networks, prioritized by RSSI

Whenever the manager fails to connect, the network in question will be
blocked/deprioritized for 5 minutes. This also happens when the user
initiates the disconnect, and in that case the manager will not try
to reconnect to any network (since the decision to disconnect was
intentional).

When a connection is established, the list of blocked networks is
cleared and the manager will not try to find another network.

The task is suspended when not trying to reconnect.

There is room for improvement here - mainly in the way networks are
prioritized and blocked - but this is good enough for our purposes
(for now, at least).